refactor: consolidate utility modules into single common module
This commit consolidates multiple utility modules (helpers, json_utils, yaml_parser, ingestion_utils) into a single utils.common module to reduce file count and simplify imports across the codebase. All imports have been updated to reference the new consolidated module.
