+++ data_access/cypher_builders/native_builders.py
@@ -7,7 +7,7 @@ Eliminates the intermediate dict serialization layer for performance optimizatio
 from typing import TYPE_CHECKING, Any
 
 from processing.entity_deduplication import generate_entity_id
-from utils.helpers import flatten_dict
+from utils.common import flatten_dict
 
 if TYPE_CHECKING:
     from models.kg_models import CharacterProfile, WorldItem
