+++ orchestration/nana_orchestrator.py
@@ -43,7 +43,7 @@ from processing.revision_logic import revise_chapter_draft_logic
 from processing.text_deduplicator import TextDeduplicator
 from processing.zero_copy_context_generator import ZeroCopyContextGenerator
 from ui.rich_display import RichDisplayManager
-from utils.ingestion_utils import split_text_into_chapters
+from utils.common import split_text_into_chapters
 
 try:
     from rich.logging import RichHandler
