+++ tests/test_yaml_parser.py
@@ -4,7 +4,7 @@ import unittest
 
 import yaml  # For creating test files
 
-from utils.yaml_parser import (
+from utils.common import (
     load_yaml_file,
     normalize_keys_recursive,
 )  # Assuming yaml_parser is in PYTHONPATH
